+++ src/edgelist.c
@@ -9,18 +9,17 @@
 // 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 void ELinitialize(context_t *ctx) {
-  int i;
   freeinit(&ctx->hfl, sizeof **ctx->ELhash);
   ctx->ELhashsize = 2 * ctx->sqrt_nsites;
   ctx->ELhash = (struct Halfedge **)myalloc(ctx, sizeof *ctx->ELhash * ctx->ELhashsize);
-  for (i = 0; i < ctx->ELhashsize; i += 1)
-    ctx->ELhash[i] = (struct Halfedge *)NULL;
-  ctx->ELleftend = HEcreate(ctx, (struct Edge *)NULL, 0);
-  ctx->ELrightend = HEcreate(ctx, (struct Edge *)NULL, 0);
-  ctx->ELleftend->ELleft = (struct Halfedge *)NULL;
+  for (int i = 0; i < ctx->ELhashsize; i++)
+    ctx->ELhash[i] = NULL;
+  ctx->ELleftend = HEcreate(ctx, NULL, 0);
+  ctx->ELrightend = HEcreate(ctx, NULL, 0);
+  ctx->ELleftend->ELleft = NULL;
   ctx->ELleftend->ELright = ctx->ELrightend;
   ctx->ELrightend->ELleft = ctx->ELleftend;
-  ctx->ELrightend->ELright = (struct Halfedge *)NULL;
+  ctx->ELrightend->ELright = NULL;
   ctx->ELhash[0] = ctx->ELleftend;
   ctx->ELhash[ctx->ELhashsize - 1] = ctx->ELrightend;
 }
@@ -34,8 +33,8 @@ struct Halfedge *HEcreate(context_t *ctx, struct Edge *e, int pm) {
   answer = (struct Halfedge *)getfree(ctx, &ctx->hfl);
   answer->ELedge = e;
   answer->ELpm = pm;
-  answer->PQnext = (struct Halfedge *)NULL;
-  answer->vertex = (struct Site *)NULL;
+  answer->PQnext = NULL;
+  answer->vertex = NULL;
   answer->ELrefcnt = 0;
   return (answer);
 }
@@ -59,16 +58,16 @@ struct Halfedge *ELgethash(context_t *ctx, int b) {
   struct Halfedge *he;
 
   if (b < 0 || b >= ctx->ELhashsize)
-    return ((struct Halfedge *)NULL);
+    return (NULL);
   he = ctx->ELhash[b];
-  if (he == (struct Halfedge *)NULL || he->ELedge != (struct Edge *)DELETED)
+  if (he == NULL || he->ELedge != (struct Edge *)DELETED)
     return (he);
 
   /* Hash table points to deleted half edge.  Patch as necessary. */
-  ctx->ELhash[b] = (struct Halfedge *)NULL;
-  if ((he->ELrefcnt -= 1) == 0)
+  ctx->ELhash[b] = NULL;
+  if ((he->ELrefcnt--) == 0)
     makefree((struct Freenode *)he, (struct Freelist *)&ctx->hfl);
-  return ((struct Halfedge *)NULL);
+  return (NULL);
 }
 
 
@@ -76,21 +75,18 @@ struct Halfedge *ELgethash(context_t *ctx, int b) {
 // 
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 struct Halfedge *ELleftbnd(context_t *ctx, struct Point *p) {
-  int i, bucket;
   struct Halfedge *he;
 
   /* Use hash table to get close to desired halfedge */
-  bucket = (p->x - ctx->xmin) / ctx->deltax * ctx->ELhashsize;
+  int bucket = (p->x - ctx->xmin) / ctx->deltax * ctx->ELhashsize;
   if (bucket < 0)
     bucket = 0;
   if (bucket >= ctx->ELhashsize)
     bucket = ctx->ELhashsize - 1;
   he = ELgethash(ctx, bucket);
-  if (he == (struct Halfedge *)NULL) {
-    for (i = 1; 1; i += 1) {
-      if ((he = ELgethash(ctx, bucket - i)) != (struct Halfedge *)NULL)
-        break;
-      if ((he = ELgethash(ctx, bucket + i)) != (struct Halfedge *)NULL)
+  if (he == NULL) {
+    for (int i = 1; 1; i++) {
+      if ( ((he = ELgethash(ctx, bucket - i)) != NULL) || ((he = ELgethash(ctx, bucket + i)) != NULL) )
         break;
     };
   };
@@ -107,10 +103,10 @@ struct Halfedge *ELleftbnd(context_t *ctx, struct Point *p) {
 
   /* Update hash table and reference counts */
   if (bucket > 0 && bucket < ctx->ELhashsize - 1) {
-    if (ctx->ELhash[bucket] != (struct Halfedge *)NULL)
-      ctx->ELhash[bucket]->ELrefcnt -= 1;
+    if (ctx->ELhash[bucket] != NULL)
+      ctx->ELhash[bucket]->ELrefcnt--;
     ctx->ELhash[bucket] = he;
-    ctx->ELhash[bucket]->ELrefcnt += 1;
+    ctx->ELhash[bucket]->ELrefcnt++;
   };
   return (he);
 }
@@ -137,14 +133,14 @@ struct Halfedge *ELleft(struct Halfedge *he) {
 
 
 struct Site *leftreg(context_t *ctx, struct Halfedge *he) {
-  if (he->ELedge == (struct Edge *)NULL)
+  if (he->ELedge == NULL)
     return (ctx->bottomsite);
   return (he->ELpm == le ? he->ELedge->reg[le] : he->ELedge->reg[re]);
 }
 
 
 struct Site *rightreg(context_t *ctx, struct Halfedge *he) {
-  if (he->ELedge == (struct Edge *)NULL)
+  if (he->ELedge == NULL)
     return (ctx->bottomsite);
   return (he->ELpm == le ? he->ELedge->reg[re] : he->ELedge->reg[le]);
 }