+++ NEWS.md
@@ -1,3 +1,10 @@
+
+# rvoronoi 0.1.0.9000  2024-08-30
+
+* Swtich from `float` to `double` in C code
+* Some general cosmetic C changes
+
+
 # rvoronoi 0.1.0  2024-08-29
 
 * Initial release